requirements.txt
filelock==3.18.0
fsspec==2025.5.1
Jinja2==3.1.6
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.5
numpy==2.3.0
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.2.5
nvidia-nvtx-cu12==12.8.90
pillow==11.2.1
pytorch-triton==3.3.1+gitc8757738
setuptools==78.1.1
shapely==2.1.1
sympy==1.14.0
torch==2.8.0.dev20250623+cu128
torchaudio==2.8.0.dev20250623+cu128
torchvision==0.23.0.dev20250623+cu128
tqdm==4.67.1
typing_extensions==4.14.0
wheel==0.45.1
